add test to verify cgas is min(d, cgas) instead of min(d, ggas)
